A first working prototype which allows local file transfers
A very basic proof of concept that can find a server on the local
network (finds the first one to respond) and can transfer a file to it
if it's the first to respond.

Right now, only the fastest UDP discover is considered so this pretty
much only works with one server listening on the local network. The
idea is to have a working path before putting in more work.