Listen for multiple responses when discovering servers
We only took the first response that we received when discovering
carbon servers. This commit listens for multiple responses and stops
listenning when it hasn't received a response for a second.

It does this by returning an iterator instead of a list of hosts. This
iterator blocks for 1 second at most per response when querying. This
allows us handle responses as they come in.